Serve static files with arbitrary epsg crs
