+++ src/config.py
@@ -142,6 +142,9 @@ class EnvVariable:  # pylint: disable=too-few-public-methods
 
     POSTGRES_HOST = _get_env_variable("POSTGRES_HOST", default="localhost")
     POSTGRES_PORT = _get_env_variable("POSTGRES_PORT", default="5431")
+    POSTGRES_INTERNAL_HOST = "db_postgres" \
+        if _get_env_variable("POSTGRES_HOST", default="localhost") == "localhost" \
+        else _get_env_variable("POSTGRES_HOST")
     POSTGRES_DB = _get_env_variable("POSTGRES_DB", default="db")
     POSTGRES_USER = _get_env_variable("POSTGRES_USER", default="postgres")
     POSTGRES_PASSWORD = _get_env_variable("POSTGRES_PASSWORD")
