+++ src/geoserver/database_layers.py
@@ -1,16 +1,71 @@
 # -*- coding: utf-8 -*-
+# Copyright © 2021-2025 Geospatial Research Institute Toi Hangarau
+# LICENSE: https://github.com/GeospatialResearch/Digital-Twins/blob/master/LICENSE
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 """Functions to handle serving database layers and views via geoserver."""
 
 import logging
 from http import HTTPStatus
 
+import geopandas as gpd
 import requests
 
 from src.config import EnvVariable
+from src.digitaltwin.data_to_db import check_table_exists
+from src.digitaltwin.setup_environment import get_database
 from src.geoserver.geoserver_common import create_workspace_if_not_exists, get_geoserver_url
 
 log = logging.getLogger(__name__)
 _xml_header = {"Content-type": "text/xml"}
+
+MAIN_DB_STORE_NAME = f"{EnvVariable.POSTGRES_DB} PostGIS"
+
+
+def get_workspace_layers(workspace_name: str, data_store_name: str = MAIN_DB_STORE_NAME):
+    """
+    Retrieve all layer names from a geoserver workspace.
+
+    Parameters
+    ----------
+    workspace_name : str
+        The name of the geoserver workspace being queried.
+    data_store_name : str = src.geoserver.database_layers.MAIN_DB_STORE_NAME
+        The name of the geoserver data store to query.
+
+    Returns
+    -------
+    list[str]
+        The names of each layer, not including the workspace name.
+
+    Raises
+    -------
+    HTTPError
+        If geoserver responds with anything but OK, raises it as an exception since it is unexpected.
+    """
+    db_exists_response = requests.get(
+        f'{get_geoserver_url()}/workspaces/{workspace_name}/datastores/{data_store_name}/featuretypes.json',
+        auth=(EnvVariable.GEOSERVER_ADMIN_NAME, EnvVariable.GEOSERVER_ADMIN_PASSWORD)
+    )
+    response_data = db_exists_response.json()
+    # Parse JSON structure to get list of feature names
+    top_layer_node = response_data["featureTypes"]
+    # defaults to empty list if no layers exist
+    layers = top_layer_node["featureType"] if top_layer_node else []
+    layer_names = [layer["name"] for layer in layers]
+    return layer_names
 
 
 def create_datastore_layer(workspace_name: str, data_store_name: str, layer_name: str, metadata_elem: str = "") -> None:
@@ -38,38 +93,39 @@ def create_datastore_layer(workspace_name: str, data_store_name: str, layer_name
     """
     layer_full_name = f"{workspace_name}:{layer_name}"
     log.info(f"Creating datastore layer '{layer_full_name}' if it does not already exist.")
-    db_exists_response = requests.get(
-        f'{get_geoserver_url()}/workspaces/{workspace_name}/datastores/{data_store_name}/featuretypes.json',
-        auth=(EnvVariable.GEOSERVER_ADMIN_NAME, EnvVariable.GEOSERVER_ADMIN_PASSWORD)
-    )
-    response_data = db_exists_response.json()
-    # Parse JSON structure to get list of feature names
-    top_layer_node = response_data["featureTypes"]
-    # defaults to empty list if no layers exist
-    layers = top_layer_node["featureType"] if top_layer_node else []
-    layer_names = [layer["name"] for layer in layers]
-    if layer_name in layer_names:
+
+    if layer_name in get_workspace_layers(workspace_name):
         # If the layer already exists, we don't have to add it again, and can instead return
         log.debug(f"Datastore layer '{layer_full_name}' already exists.")
         return
+    # Find SRS/CRS information
+    engine = get_database()
+    if check_table_exists(engine, layer_name):
+        gdf = gpd.read_postgis(f'SELECT * FROM "{layer_name}"', engine, "geometry")
+    else:
+        # Default values if nothing else is available
+        gdf = gpd.read_file("selected_polygon.geojson")
+    minx, miny, maxx, maxy = gdf.total_bounds
+    minx4326, miny4326, maxx4326, maxy4326 = gdf.to_crs(4326).total_bounds
+    crs = gdf.crs.to_epsg()
     # Construct new layer request
     data = f"""
         <featureType>
             <name>{layer_name}</name>
             <title>{layer_name}</title>
-            <srs>EPSG:2193</srs>
+            <srs>EPSG:{crs}</srs>
             <nativeBoundingBox>
-                <minx>1563837.8771000002</minx>
-                <maxx>5175183.3933</maxx>
-                <miny>1580158.7676</miny>
-                <maxy>5185241.6301</maxy>
-                <crs class="projected">EPSG:2193</crs>
+                <minx>{minx}</minx>
+                <maxx>{maxx}</maxx>
+                <miny>{miny}</miny>
+                <maxy>{maxx}</maxy>
+                <crs class="projected">EPSG:{crs}</crs>
             </nativeBoundingBox>
             <latLonBoundingBox>
-                <minx>174.9216900648596</minx>
-                <maxx>175.24540158384417</maxx>
-                <miny>-40.96195378261949</miny>
-                <maxy>-40.79359007336633</maxy>
+                <minx>{minx4326}</minx>
+                <maxx>{maxx4326}</maxx>
+                <miny>{miny4326}</miny>
+                <maxy>{maxy4326}</maxy>
                 <crs>EPSG:4326</crs>
             </latLonBoundingBox>
             <store>
@@ -140,7 +196,7 @@ def create_db_store_if_not_exists(db_name: str, workspace_name: str, new_data_st
         <dataStore>
           <name>{new_data_store_name}</name>
           <connectionParameters>
-            <host>{EnvVariable.POSTGRES_HOST}</host>
+            <host>{EnvVariable.POSTGRES_INTERNAL_HOST}</host>
             <port>{EnvVariable.POSTGRES_PORT}</port>
             <database>{db_name}</database>
             <user>{EnvVariable.POSTGRES_USER}</user>
@@ -186,12 +242,10 @@ def create_main_db_store(workspace_name: str) -> str:
     HTTPError
        If geoserver responds with an error, raises it as an exception since it is unexpected.
     """
-    log.debug("Creating building database views if they do not exist")
-    db_name = EnvVariable.POSTGRES_DB
+    log.debug(f"Creating {MAIN_DB_STORE_NAME} store if it does not exist")
     # Create workspace if it doesn't exist, so that the namespaces can be separated if multiple dbs are running
     create_workspace_if_not_exists(workspace_name)
     # Create a new database store if geoserver is not yet configured for that database
-    data_store_name = f"{db_name} PostGIS"
-    create_db_store_if_not_exists(db_name, workspace_name, data_store_name)
-
-    return data_store_name
+    db_name = EnvVariable.POSTGRES_DB
+    create_db_store_if_not_exists(db_name, workspace_name, MAIN_DB_STORE_NAME)
+    return MAIN_DB_STORE_NAME
